phase4_routing/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
shapely>=2.0.0
boto3>=1.28.0
psycopg2-binary>=2.9.0
pytest>=7.4.0
